Compressing is now angle-limited
During compressing, the angle between segments is checked and compared
with a threshold. If it is bigger, compressing is cut short.

+++ common/c_planner_v5/EnableDebugLog.m
@@ -1,6 +1,6 @@
 function EnableDebugLog(cfg)
 global DebugConfig
 
-    DebugConfig = bitset(DebugConfig, cfg);  
+    DebugConfig = bitset(DebugConfig, int32(cfg));  
 
 end
